refactor: unify components into a single `AccountSelector`
