+++ packages/shared/components/inputs/index.js
@@ -1,3 +1,4 @@
+export { default as AccountSelector } from './AccountSelector.svelte'
 export { default as AliasInput } from './AliasInput.svelte'
 export { default as AmountInput } from './AmountInput.svelte'
 export { default as AssetAmountInput } from './AssetAmountInput.svelte'
